remove test modules not needed (or not needed yet) and refactors test_request
