--- tests/test_garbage.py
@@ -1,32 +1,0 @@
-import unittest
-from .. modules.garbage import GC, STWGC, GCI
-
-class TestGCI(unittest.TestCase):
-
-    @classmethod
-    def setUp(self):
-        pass
-
-    def test_algumaCoisa(self):
-        pass
-
-class TestGC(unittest.TestCase):
-
-    @classmethod
-    def setUp(self):
-        pass
-
-    def test_algumaCoisa(self):
-        pass
-
-class TestSTWGC(unittest.TestCase):
-
-    @classmethod
-    def setUp(self):
-        pass
-
-    def test_algumaCoisa(self):
-        pass
-
-if __name__ == '__main__':
-    unittest.main()
